Adding TokenFilter support:
This adds the "LookupAttribute" used by the Tagger to decide if a Token consumed via TokenStream#incrementToken() needs to be looked up in the FST corpus. Analyzer chains may define a TokenFilter component that does set this attribute to TRUE or FALER. Typically this will be done based on NLP (Natural Language Processing) results.

The Tagger uses the LookupAttribute as follows:

* Tags are only created for Tokens where the LookupAttribute is TRUE
* Advancing tags are continued regardless of the LookupAttribute value